Initial unit test for Rollover
Basically a minimal translation of Ryan's example script that exercises
Rollover into the unittest framework. We will want to streamline, refactor
and grow this, but it serves as an initial example.

Also moved all the existing contents of tests/ to _obsolete/tests/ since
they were not running properly and Ryan did not think they were worth
salvaging at this time.