Basic loading from postgres
Currently this will fail due to a bad query when loading
ParasiticEnergy; that is being fixed on #19. In the meantime, I
have tested that we can do a full load from postgres with that
query fixed (though the model will subsequently generate
ParasiticEnergy-related exceptions if you try to run it.)

There is also some leftover cruft from sqlite like the db_path
argument to Config.init_db() and hard-coded postgres connection
parameters (user='michael') that will need to be replaced as we
complete the transition.